addProfilePageClient and hydration page
