+++ src/lib/utils/formatUserName.ts
