+++ src/lib/utils/getQueryClient.ts
@@ -1,0 +1,5 @@
+import { cache } from "react";
+import { QueryClient } from "@tanstack/react-query";
+
+const getQueryClient = cache(() => new QueryClient()) as () => QueryClient;
+export default getQueryClient;
